// functions/portal/src/types.d.ts
interface IPortalPlugin {
  name: string
  featureFlag?: string
  enabled?: string
  children?: IPortalPlugin[]
}

export interface PaConfig {
  meta: {
    configId: string
    configName: string
  }
}

export type PaConfigType = `${PA_CONFIG_TYPE}`
export interface IMetadataConfigTag {
  name: string
}

export interface IMetadataConfigAttribute {
  id: string
  name: string
  category: ATTRIBUTE_CONFIG_CATEGORIES
  dataType: ATTRIBUTE_CONFIG_DATA_TYPES
  isDisplayed: boolean
}
export interface IFeatureDto {
  feature: string
  isEnabled: boolean
}

export interface IFeatureUpdateDto {
  features: IFeatureDto[]
}